refactor(itinerary components): Remove unused props.
