+++ lib/components/narrative/default/default-itinerary.js
@@ -308,7 +308,6 @@ class DefaultItinerary extends NarrativeItinerary {
               itinerary={itinerary}
               LegIcon={LegIcon}
               setActiveLeg={setActiveLeg}
-              timeOptions={timeOptions}
             />
           </>
         )}
