+++ lib/components/narrative/line-itin/line-itinerary.js
@@ -60,7 +60,6 @@ export default class LineItinerary extends NarrativeItinerary {
             // (will cause error when clicking on itinerary summary).
             // Use the one passed by NarrativeItineraries instead.
             setActiveLeg={setActiveLeg}
-            timeOptions={timeOptions}
           />
         ) : null}
         {ItineraryFooter && <ItineraryFooter />}
